requirements.txt
python-gitlab
python-dotenv
neo4j
ruamel.yaml
cwl-utils
antlr4-python3-runtime